refactor: move to components to providers
