+++ .husky/pre-commit
@@ -1,2 +1,1 @@
 npx lint-staged
-npx lint-staged
