+++ package.json
@@ -7,6 +7,7 @@
     "dev": "vite",
     "build": "tsc -b && vite build",
     "lint": "eslint --cache 'src/**/*.{ts,tsx}'",
+    "tsc": "tsc --noEmit",
     "format": "prettier --write --cache 'src/**/*.{ts,tsx}'",
     "lint-staged": "lint-staged",
     "preview": "vite preview",
@@ -18,15 +19,14 @@
   "lint-staged": {
     "**/*.{tsx,ts,jsx,js}": [
       "eslint --fix --cache",
-      "prettier --write --cache"
-    ],
-    "*.js": "eslint --cache --fix",
-    "*.{ts,tsx}": "prettier --write"
+      "prettier --write --cache"      
+    ]
   },
   "dependencies": {
     "@emotion/babel-plugin": "^11.12.0",
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
+    "@emotion/css": "^11.13.0",
     "@tanstack/react-query": "^5.56.2",
     "csstype": "^3.1.3",
     "react": "^18.3.1",
@@ -35,7 +35,6 @@
   },
   "devDependencies": {
     "@chromatic-com/storybook": "^2.0.2",
-    "@emotion/css": "^11.13.0",
     "@eslint/js": "^9.9.0",
     "@storybook/addon-essentials": "^8.3.0",
     "@storybook/addon-interactions": "^8.3.0",
