+++ src/components/providers/index.tsx
@@ -1,0 +1,6 @@
+import { ReactNode } from 'react';
+import GlobalStylesProvider from './GlobalStylesProvider/index.provider';
+
+export default function AppProviders({ children }: { children: ReactNode }) {
+  return <GlobalStylesProvider>{children}</GlobalStylesProvider>;
+}
