+++ src/main.tsx
@@ -1,9 +1,9 @@
 import { createRoot } from 'react-dom/client';
 import App from './App.tsx';
-import StyleProvider from '@assets/styles/StyleProvider';
+import AppProviders from '@components/providers/index.tsx';
 
 createRoot(document.getElementById('root')!).render(
-  <StyleProvider>
+  <AppProviders>
     <App />
-  </StyleProvider>,
+  </AppProviders>,
 );
